Starting to reorganize and refactor
